--- a/core/protocols/aws-query-protocol/src/main/java/software/amazon/awssdk/protocols/query/internal/marshall/ListQueryMarshaller.java
+++ b/core/protocols/aws-query-protocol/src/main/java/software/amazon/awssdk/protocols/query/internal/marshall/ListQueryMarshaller.java
@@ -19,6 +19,7 @@ import java.util.List;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.core.SdkField;
 import software.amazon.awssdk.core.traits.ListTrait;
+import software.amazon.awssdk.core.traits.TraitType;
 import software.amazon.awssdk.core.util.SdkAutoConstructList;
 
 /**
@@ -52,7 +53,7 @@ public class ListQueryMarshaller implements QueryMarshaller<List<?>> {
             return;
         }
         for (int i = 0; i < val.size(); i++) {
-            ListTrait listTrait = sdkField.getTrait(ListTrait.class);
+            ListTrait listTrait = sdkField.getTrait(ListTrait.class, TraitType.LIST_TRAIT);
             String listPath = pathResolver.resolve(path, i, listTrait);
             QueryMarshaller<Object> marshaller = context.marshallerRegistry().getMarshaller(
                 ((SdkField<?>) listTrait.memberFieldInfo()).marshallingType(), val);