--- a/core/protocols/protocol-core/src/main/java/software/amazon/awssdk/protocols/core/StringToInstant.java
+++ b/core/protocols/protocol-core/src/main/java/software/amazon/awssdk/protocols/core/StringToInstant.java
@@ -23,6 +23,7 @@ import software.amazon.awssdk.core.SdkField;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.core.protocol.MarshallLocation;
 import software.amazon.awssdk.core.traits.TimestampFormatTrait;
+import software.amazon.awssdk.core.traits.TraitType;
 import software.amazon.awssdk.utils.DateUtils;
 
 /**
@@ -80,7 +81,7 @@ public final class StringToInstant implements StringToValueConverter.StringToVal
     }
 
     private TimestampFormatTrait.Format resolveTimestampFormat(SdkField<Instant> field) {
-        TimestampFormatTrait trait = field.getTrait(TimestampFormatTrait.class);
+        TimestampFormatTrait trait = field.getTrait(TimestampFormatTrait.class, TraitType.TIMESTAMP_FORMAT_TRAIT);
         if (trait == null) {
             TimestampFormatTrait.Format format = defaultFormats.get(field.location());
             if (format == null) {