--- a/core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/DataTypeConversionFailureHandlingTrait.java
+++ b/core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/DataTypeConversionFailureHandlingTrait.java
@@ -19,4 +19,8 @@ import software.amazon.awssdk.annotations.SdkProtectedApi;
 
 @SdkProtectedApi
 public class DataTypeConversionFailureHandlingTrait implements Trait {
+    @Override
+    public TraitType type() {
+        return TraitType.DATA_TYPE_CONVERSION_FAILURE_HANDLING_TRAIT;
+    }
 }