--- a/core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/DefaultValueTrait.java
+++ b/core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/DefaultValueTrait.java
@@ -62,4 +62,9 @@ public final class DefaultValueTrait implements Trait {
     public static DefaultValueTrait idempotencyToken() {
         return new DefaultValueTrait(IdempotentUtils.getGenerator());
     }
+
+    @Override
+    public TraitType type() {
+        return TraitType.DEFAULT_VALUE_TRAIT;
+    }
 }