… core/protocols/protocol-core/src/main/java/software/amazon/awssdk/protocols/core/AbstractMarshallingRegistry.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/SdkField.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/protocol/MarshallingKnownType.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/protocol/MarshallingType.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/JsonValueTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/ListTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/LocationTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/MapTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/PayloadTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/RequiredTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/TimestampFormatTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/Trait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/TraitType.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/XmlAttributeTrait.java
… core/sdk-core/src/main/java/software/amazon/awssdk/core/traits/XmlAttributesTrait.java
… services-custom/s3-transfer-manager/src/main/java/software/amazon/awssdk/transfer/s3/internal/serialization/TransferManagerJsonUnmarshaller.java