<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context

Improvement to the lookup performance of marshallers and traits. 

Both traits and marshalling types are open classes but with well known, finite, and, statically defined instances. We can improve the performance of the lookups of these values by enumerating them using an enum and associating the enum values to the concrete types, and then using a two level lookup: the fast level 1 uses the an enum map and the associate enumeration value as key, while the l2 uses a regular map with the values as keys as they are. This change improves not only RPCv2, but all the other protocols as well.

This change aims to reduce, or remove the hump highlighted in the flamegraph below


<img width="1021" alt="before-highligthed" src="https://github.com/user-attachments/assets/9339ae33-758f-425f-b39b-b7327bdc2853">


**Before**

```
Benchmark                              (protocol)  (size)  Mode  Cnt      Score     Error  Units
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2   small  avgt    5   3415.127 ±  64.585  ns/op
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2  medium  avgt    5   5668.994 ±  16.272  ns/op
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2     big  avgt    5  19475.973 ± 133.802  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json   small  avgt    5  10150.744 ±  28.942  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json  medium  avgt    5  21037.573 ± 139.545  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json     big  avgt    5  86466.945 ± 709.876  ns/op
```

And zoomed in flamegraph

![before-enum-map](https://github.com/user-attachments/assets/4e0016e6-0ab7-43a1-a116-999467f0008d)


**After**

```
Benchmark                              (protocol)  (size)  Mode  Cnt      Score      Error  Units
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2   small  avgt    5   3166.243 ±  151.799  ns/op
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2  medium  avgt    5   5349.115 ±  172.858  ns/op
JsonMarshallerBenchmark.unmarshall  smithy-rpc-v2     big  avgt    5  17517.889 ±   41.898  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json   small  avgt    5  10178.752 ±  167.762  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json  medium  avgt    5  20440.322 ±  223.031  ns/op
JsonMarshallerBenchmark.unmarshall       aws-json     big  avgt    5  83252.862 ± 1159.517  ns/op
```

And zoomed in flamegraph

![after-enum-map](https://github.com/user-attachments/assets/1aa1d297-8905-401f-9dab-e9c00b12fb8f)


Full flamegraph after this change

<img width="1022" alt="after-full" src="https://github.com/user-attachments/assets/7834c674-be27-4568-99d7-590631d88077">


## Modifications
<!--- Describe your changes in detail -->

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate)

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

## Checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] I have read the [CONTRIBUTING](https://github.com/aws/aws-sdk-java-v2/blob/master/CONTRIBUTING.md) document
- [ ] Local run of `mvn install` succeeds
- [ ] My code follows the code style of this project
- [ ] My change requires a change to the Javadoc documentation
- [ ] I have updated the Javadoc documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] All new and existing tests passed
- [ ] I have added a changelog entry. Adding a new entry must be accomplished by running the `scripts/new-change` script and following the instructions. Commit the new file created by the script in `.changes/next-release` with your changes.
- [ ] My change is to implement 1.11 parity feature and I have updated [LaunchChangelog](https://github.com/aws/aws-sdk-java-v2/blob/master/docs/LaunchChangelog.md)

## License
<!--- The SDK is released under the Apache 2.0 license (http://aws.amazon.com/apache2.0/), so any code you submit will be released under that license -->
<!--- For substantial contributions, we may ask you to sign a Contributor License Agreement (http://en.wikipedia.org/wiki/Contributor_License_Agreement) -->
<!--- Put an `x` in the below box if you confirm that this request can be released under the Apache 2 license -->
- [ ] I confirm that this pull request can be released under the Apache 2 license